requirements.txt

# ----------------------
# Flask
# ----------------------
flask

# ----------------------
# Cors
# ----------------------
flask_cors

# ----------------------
# Testing
# ----------------------
# mock


# ----------------------
# Production Server
# ----------------------
gunicorn

# ----------------------
# Utils
# ----------------------
bson
pymongo
PyJWT
bcrypt
datetime
stripe
dnspython
